satellite/requirements.txt
RPi.GPIO
pygame
requests
